+++ cmd/todos.go
@@ -1,15 +1,11 @@
-package main
+package todos
 
 import (
 	"errors"
 	"fmt"
-	"log"
 	"os"
-	"path/filepath"
 	"strconv"
 	"time"
-	"todo/cli"
-	"todo/storage"
 
 	"github.com/aquasecurity/table"
 )
@@ -106,21 +102,4 @@ func (todos *Todos) Print() {
 	}
 
 	table.Render()
-}
-
-func main() {
-	todos_storage_path := os.Getenv("TODOS_STORAGE")
-	if todos_storage_path == "" {
-		home, err := os.UserHomeDir()
-		if err != nil {
-			log.Fatal(err)
-		}
-		todos_storage_path = filepath.Join(home, ".todos.json")
-	}
-	todos_storage := storage.NewStorage[Todos](todos_storage_path)
-	todos := &Todos{}
-	todos_storage.Load(todos)
-	cmdFlags := cli.NewCmdFlags()
-	cmdFlags.Execute(todos)
-	todos_storage.Save(*todos)
 }
